chore: namespcae, folder structure
